requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.5
django-autoslug==1.9.8
django-ckeditor==6.0.0
django-etc==1.3.0
django-heroku==0.3.1
django-hitcount==1.3.3
django-js-asset==1.2.2
gunicorn==20.0.4
Pillow==8.1.0
psycopg2==2.8.6
pytz==2020.5
sqlparse==0.4.1
whitenoise==5.2.0
